feat: implement TagManagementViewController
#15
